# .gnva-sauvegardes/rapports-2026-10-04T05-53-07-902Z/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: API GNVA, autorisations territoriales et opérations transactionnelles
servers:
  - url: /api
paths:
...
  /v1/auth/mot-de-passe:
    post:
      operationId: changerMotDePasse
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ChangementMotDePasse" }
      responses:
        "200":
          description: Modifié
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Resultat" }
...
components:
  schemas:
...
    Resultat:
      type: object
      required: [message]
      properties:
        message: { type: string }
        nombre: { type: integer }
...
    ChangementMotDePasse:
      type: object
      required: [actuel, nouveau]
      properties:
        actuel: { type: string }
        nouveau: { type: string, minLength: 12 }
...